refract: clear code remove useless folder
