+++ libs/ui/pages/settings/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useOrganisationSettingsPageLogic } from './logic';
 import Page from 'ui/components/contents/page';
-import { useSidebarPageLogic } from 'ui/components/contents/sidebar-page/logic';
-import SidebarPage from 'ui/components/contents/sidebar-page';
 import OrganisationSettings from '../../components/contents/organisation-settings/index';
+import SideBar from '../../components/sidebar/index';
+import { useSidebarLogic } from '../../components/sidebar/logic';
 
 const OrganisationSettingsPage = () => {
     // Attributes
@@ -18,26 +18,42 @@ const OrganisationSettingsPage = () => {
         handleSwitchOrgansiation,
         actualOrganisationUser,
         organisationUserData,
-        organisationProjectData,
         filterProjectValue,
         setFilterProjectValue,
         activeProject,
         setActiveProject,
-    } = useSidebarPageLogic();
+        createProjectModalDisclosure,
+        isDisableOnCloseProjectModal,
+        handleToggleIsOrganisationClicked,
+        handleOnOptionClick,
+        isOrganisationClicked,
+        options,
+        activeOptionKey,
+        setIsOrganisationClicked,
+    } = useSidebarLogic();
     // Render
     return (
         <Page>
-            <SidebarPage
+            <SideBar
+                handleToggleIsOrganisationClicked={
+                    handleToggleIsOrganisationClicked
+                }
+                handleOnOptionClick={handleOnOptionClick}
+                isOrganisationClicked={isOrganisationClicked}
+                options={options}
+                activeOptionKey={activeOptionKey}
+                setIsOrganisationClicked={setIsOrganisationClicked}
+                createProjectModalDisclosure={createProjectModalDisclosure}
                 handleOnCreateProject={handleOnCreateProject}
                 handleOnCreateOrganisation={handleOnCreateOrganisation}
                 handleSwitchOrgansiation={handleSwitchOrgansiation}
                 actualOrganisationUser={actualOrganisationUser}
                 organisationUserData={organisationUserData}
-                organisationProjectData={organisationProjectData}
                 filterProjectValue={filterProjectValue}
                 setFilterProjectValue={setFilterProjectValue}
                 activeProject={activeProject}
                 setActiveProject={setActiveProject}
+                isDisableOnCloseProjectModal={isDisableOnCloseProjectModal}
             />
             <OrganisationSettings
                 handleOnDeleteOrganisation={handleOnDeleteOrganisation}
